PAT: fix clang-tidy findings and a test leak
- pat.c: (void) the feature-name snprintf calls, store letter
  multipliers as int instead of int8_t, uppercase literal suffixes, and
  check overlay alongside run_has_fresh_tile (a fresh tile implies an
  overlay; the analyzer could not see it).
- config.c: move the PAT options' help into add_pat_help_arg, which
  brings add_help_arg_to_string_builder back under the size threshold;
  include kwg.h directly.
- pat_gen.c: replace a nested conditional with if/else.
- Include what is used in player.c, players_data.c and autoplay.c, and
  cast a uint64_t to double explicitly in autoplay_results.c.
- Tests: free the zeroed PAT in test_pat_movegen_integration once
  players_data no longer holds it (LeakSanitizer in CI), initialize
  zero_weight_equities, enums instead of macros and string_to_int instead
  of atoi in pat_opening_sim_test.c, direct includes and explicit
  conversions in pat_test.c, and drop unused unistd.h includes.

Static CSW24 play (2,000 pairs, -pat CSW24) is byte-identical to before.

+++ src/ent/autoplay_results.c
@@ -554,7 +554,8 @@ char *game_data_human_readable_str(const GameData *gd, bool divergent,
                                 : 0.0;
       string_builder_add_formatted_string(
           sb, "  %d tiles: mean %.3f, SE %.3f (%llu games, %.1f%%)\n", tiles,
-          mean, se, (unsigned long long)n, 100.0 * (double)n / opening_samples);
+          mean, se, (unsigned long long)n,
+          100.0 * (double)n / (double)opening_samples);
     }
   }
   string_builder_add_string(sb, "\n");
